requirements.txt
# pip install -U -r requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
pillow~=10.0.0
easydict~=1.10
# pycocotools>=2.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
#torch>=1.6.0
#torchvision>=0.7.0
tqdm>=4.41.0

torch~=2.0.1
torchvision~=0.15.2
Django~=4.2.4
djangorestframework~=3.14.0
requests~=2.31.0